.pipeline/projects/drop_servicing_tool/workspace/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drop-servicing-tool"
version = "0.1.0"
description = "Execute service delivery workflows as SOPs using LLM-powered agentic workflows."
requires-python = ">=3.10"
dependencies = [
    "pyyaml",
    "typer",
    "pydantic",
]

[project.scripts]
sop = "drop_servicing_tool.cli:app"

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
include = ["drop_servicing_tool*"]
